requirements.txt
alembic==1.0.8
astroid==2.2.5
autopep8==1.4.3
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Mako==1.0.8
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.2
typed-ast==1.3.1
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
